Statistical tests for 'Counts per Volume'
After longer discussion with Fluri he helped me to implement the
relevant statistical tests to see differences in the 'Count per volumes'
in between the different animals.
It seems that we see differences for all combinations except between
days 10 and 21, as one can also see from the plot.
So it looks like we're all good.

Now we want to make this a bit prettier and use the 'itertools' package
to iteratively go through all day combinations.